tools/csv-edit/pyproject.toml
[project]
name = "csv-edit"
version = "0.1.0"
description = "A terminal-based CSV editor that integrates directly with GitHub."
requires-python = ">=3.9"
dependencies = [
    "textual>=0.52.1",
    "PyGithub>=2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0"
]

[project.scripts]
csv-edit = "csv_edit.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
